Added Ricochet Ability
Trying out a new ability to be the main mechanic for the game.:

Ricochet - Character throws a bladerang-type weapon that is launched towards the enemy. Once it hits an enemy, it will launch itself in the air where it can be caught by the player. Catching the bladerang before it hits the ground will give the player a positive buff like a stacking damage buff. Open to being other things positive statuses as well.

So far, I was able to add in the bladerang and allow the player to launch it. Still a lot of things to fix so it works properly, but the basic functionality is working.